Refactored the DAO layer and implemented the scanned images DAO. Changed the return type in Room DAOs from LiveData to Flowable to have more flexibility in scheduling database reads on threads.

+++ ReceiptScan/app/src/main/java/com/lucianbc/receiptscan/domain/model/Definitions.kt
@@ -1,5 +1,5 @@
 package com.lucianbc.receiptscan.domain.model
 
-typealias ID = Int?
+typealias ID = Long?
 typealias ScanAnnotations = List<ScanInfoBox>
 typealias Drafts = List<ReceiptDraft>
